backend/requirements.txt
accelerate
huggingface-hub
joblib
langchain
langchain-community
langchain-core
langchain-experimental
langchain-groq
langchain-text-splitters
langsmith
language_data
multidict
multipledispatch
multitasking
neo4j
nltk
numpy
packaging
pandas
pathlib
python-dotenv
sentence-transformers
torch
transformers
decorator
json-repair
tiktoken
fastapi
uvicorn
psycopg2-binary
python-magic
pymupdf4llm
python-multipart
langchain_huggingface
